requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro~=0.19.6
kedro-datasets>=3.0; python_version >= "3.9"
kedro-datasets>=1.0; python_version < "3.9"
kedro-telemetry>=0.3.1
kedro-viz>=6.7.0
notebook
ipython>=8.10
jupyterlab>=3.0
kedro~=0.19.5
kedro-telemetry>=0.3.1
notebook
ruff~=0.1.8
fsspec~=2023.10.0
numpy~=1.26.4
openai~=1.30.1
pydantic~=1.10.12
python-dotenv~=1.0.1
torch~=2.3.0
torchaudio~=2.3.0